增加缓存清理插件

+++ lib/pod/auto_switch_source_hook.rb
@@ -1,8 +1,12 @@
 # frozen_string_literal: true
 
-Pod::HooksManager.register('auto-switch-source', :pre_install) do |ctx, _|
-  project_root = "#{Pod::Config.instance.project_root.to_s}/Pods"
-  cache_root = "#{Pod::Config.instance.cache_root.to_s}/Pods"
+Pod::HooksManager.register('cocoapods-publish', :pre_install) do |ctx, _|
+  next unless ctx.podfile.plugins.keys.include?('cocoapods-publish')
+
+  puts '开始清理缓存...'.yellow
+
+  project_root = "#{Pod::Config.instance.project_root}/Pods"
+  cache_root = "#{Pod::Config.instance.cache_root}/Pods"
 
   Dir.glob("#{cache_root}/**/BT*/")
      .each { |path| `rm -rf #{path}` if Dir.exist?(path) }
@@ -10,5 +14,6 @@ Pod::HooksManager.register('auto-switch-source', :pre_install) do |ctx, _|
   Dir.glob("#{project_root}/**/BT*/")
      .each { |path| `rm -rf #{path}` if Dir.exist?(path) }
 
+  puts '缓存清理完成'.green
   next
 end
